feat(today): host the Night-detail sleep card (#today-hosted-cards)
Card 2 on the shared SleepModel foundation, mirroring card 1. Lift SleepView's
metricGrid into a standalone NightDetailCard(model:); register
HostedCard.nightDetail (byte-identical sleep.nightDetail id, existing localized
title). Reuses the Today-host model plumbing from card 1 - only adds the dispatch
case. Android NightDetailHostCard keeps the Sleep tab's per-metric detail nav via
an optional onMetricClick (Today passes none, read-only mirror).

Android compile + i18n audit green locally; Swift via gate.
