+++ android/app/src/main/java/com/noop/ui/TodayScreen.kt
@@ -3074,7 +3074,7 @@ private fun HostedCardsSection(cards: List<HostedCard>, days: List<DailyMetric>,
     // follow). Build the shared model ONCE, and only when at least one such card is actually hosted, so a
     // Today hosting none pays no cost. Same inputs + builder as the Sleep tab (buildHostedSleepModel), so
     // hosted numbers match the Sleep tab. Reused by every model-backed card. Twin of the iOS hostedSleepModel.
-    val modelBackedHosted = setOf(HostedCard.STAGES_VS_TYPICAL)
+    val modelBackedHosted = setOf(HostedCard.STAGES_VS_TYPICAL, HostedCard.NIGHT_DETAIL)
     val needsSleepModel = cards.any { it in modelBackedHosted }
     var hostedSleepModel by remember { mutableStateOf<SleepModel?>(null) }
     LaunchedEffect(needsSleepModel, days, viewModel.activeStrapId) {
@@ -3111,6 +3111,10 @@ private fun HostedCardsSection(cards: List<HostedCard>, days: List<DailyMetric>,
                 // lands — or on a device with no stage data — the model is null and the slot renders nothing
                 // this frame, matching how the Sleep tab guards the section on a null model.
                 HostedCard.STAGES_VS_TYPICAL -> hostedSleepModel?.let { StagesVsTypicalHostCard(it) }
+                // #today-hosted-cards: the Night-detail metric grid, rendered from the SAME shared SleepModel
+                // the Sleep tab uses (mirror, not copy). Null until the async build lands / no stage data —
+                // the slot renders nothing this frame, matching the Sleep tab's null-model guard.
+                HostedCard.NIGHT_DETAIL -> hostedSleepModel?.let { NightDetailHostCard(it) }
             }
         }
     }
